refactor: Repository 레이어 분리 및 패키지 구조 재구성
