+++ app/src/main/java/woowacourse/shopping/di/DataContainer.kt
@@ -1,11 +1,18 @@
 package woowacourse.shopping.di
 
-import woowacourse.shopping.repository.cart.CartRepository
-import woowacourse.shopping.repository.cart.MockCartRepository
-import woowacourse.shopping.repository.product.MockProductRepository
-import woowacourse.shopping.repository.product.ProductRepository
+import android.content.Context
+import woowacourse.shopping.domain.repository.CartRepository
+import woowacourse.shopping.data.repository.cart.MockCartRepository
+import woowacourse.shopping.data.repository.product.MockProductRepository
+import woowacourse.shopping.domain.repository.ProductRepository
 
 object DataContainer {
+    private var appContext: Context? = null
+    fun init(context: Context){
+        if(appContext==null){
+            appContext = context.applicationContext
+        }
+    }
     val cartRepository: CartRepository by lazy { MockCartRepository() }
     val productRepository: ProductRepository by lazy { MockProductRepository() }
 }
