+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
 import woowacourse.shopping.di.DataContainer
-import woowacourse.shopping.repository.cart.MockCartRepository
+import woowacourse.shopping.data.repository.cart.MockCartRepository
 
 class CartActivity : ComponentActivity() {
     @OptIn(ExperimentalMaterial3Api::class)
